Intiall Commit
A web-browser based meme game in which you have to guess the caption from the options that suits most the shown picture.  We have used python for scraping the pictures, blurring the texts in these and extraxting those texts to make options in random order. Libraries that we have used are OpenCV, ocrSpace API, imutils and Pillow.